+++ src/main/kotlin/xyz/haff/fizzbuzz/flow/FizzFlowConfiguration.kt
@@ -5,6 +5,7 @@ import org.springframework.context.annotation.Configuration
 import org.springframework.integration.dsl.StandardIntegrationFlow
 import org.springframework.integration.dsl.integrationFlow
 import org.springframework.messaging.MessageChannel
+import java.math.BigInteger
 
 @Configuration
 class FizzFlowConfiguration(
@@ -14,7 +15,7 @@ class FizzFlowConfiguration(
 
     @Bean
     fun fizzFlow(): StandardIntegrationFlow = integrationFlow(fizzChannel) {
-        transform<Long> { "Fizz" }
+        transform<BigInteger> { "Fizz" }
         channel(outputChannel)
     }
 }
